refactor(parser): split parser modules
