+++ .vscode/settings.json
@@ -5,7 +5,7 @@
             "file": ".gitmessage",
         }
     ],
-    "github.copilot.chat.agent.terminal.allowList": {
+    "chat.agent.terminal.allowList": {
         "mise run build": true,
         "mise run test": true,
         "git status": true,
